refactor : webtoon 페이지 관련된 내용을 webtoon.js로 이동
- webtoon 페이지에 관련된 코드들을 webtoon.js로 이동
