+++ public/javascripts/webtoon.js
@@ -1,0 +1,32 @@
+import { $ } from './utils.js';
+import { createGenreNav, changeClickedNavTab } from './genre-nav.js';
+import { headerNavData } from './data.js';
+
+const HEADER_NAV_CLICKED_ITEM_CLASS_NAME = 'header-nav__item--clicked';
+const HEADER_NAV_LIST_CLASS_NAME = '.header-nav__list';
+
+const WEBTOON_NAV_LIST = [
+  '홈',
+  '요일연재',
+  '웹툰',
+  '소년',
+  '드라마',
+  '로맨스',
+  '로판',
+  '액션무협',
+  'BL/GL',
+];
+
+function renderWebtoonPage(pageIndex) {
+  const contentTag = $('.content');
+  contentTag.innerText = '';
+  changeClickedNavTab(
+    headerNavData,
+    pageIndex,
+    HEADER_NAV_LIST_CLASS_NAME,
+    HEADER_NAV_CLICKED_ITEM_CLASS_NAME
+  );
+  contentTag.appendChild(createGenreNav(WEBTOON_NAV_LIST, 'webtoon-genre-nav'));
+}
+
+export { renderWebtoonPage };
